tutorial_pipeline_v2/requirements.txt
streamlit>=1.42.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
pandas>=2.2.0
google-generativeai>=0.8.3
pytest>=8.3.0
flake8>=7.1.0
mypy>=1.15.0
schedule>=1.2.2
